moved deprecated classes
